tests/wpt/mozilla/tests/mozilla/window_requestAnimationFrame.html

<html>
  <head>
    <title>Test throwing an error inside requestAnimationFrame callback</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    window.requestAnimationFrame(function() {});
    assert_equals(true, true, "rAF shouldn't explode if called before layout begins");
    </script>
  </body>
</html>
